+++ tensorcircuit/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.0.220126"
+__version__ = "0.0.220301"
 __author__ = "refraction-ray"
 
 from .cons import (